requirements.txt
# 🔥 Core App
Flask
Flask-SQLAlchemy
flask-migrate
flask-mail

flask-wtf
bcrypt
gunicorn
apscheduler

# 📧 Email + QR
pillow
qrcode[pil]

# 💳 Payments
stripe

# 🔍 Fuzzy Matching, Timezones
rapidfuzz
pytz

# 🧠 Data / AI Stack
pandas
matplotlib
duckdb
scipy
faker

# 🧠 LLM / Chatbot
ollama  # Python client for your Ollama server
anthropic>=0.25.0  # Claude API client
openai>=1.30.0     # OpenAI API client
aiohttp>=3.8.0     # Async HTTP client for AI providers
tiktoken>=0.6.0    # Token counting for OpenAI models
tenacity>=8.2.0    # Retry logic for API calls
pydantic>=2.0.0    # Data validation

premailer
requests
python-dotenv
marshmallow

# 🔒 Security
bleach  # HTML sanitization for email templates